add: more components
